--- a/RoaringBitmap/build.gradle.kts
+++ b/RoaringBitmap/build.gradle.kts
@@ -7,6 +7,7 @@ dependencies {
     testImplementation("com.google.guava:guava:${deps["guava"]}")
     testImplementation("org.apache.commons:commons-lang3:${deps["commons-lang"]}")
     testImplementation("com.esotericsoftware:kryo:5.0.0-RC1")
+    testImplementation("com.fasterxml.jackson.core", "jackson-databind", "2.10.3")
 }