--- a/RoaringBitmap/src/main/java/org/roaringbitmap/RoaringBitmap.java
+++ b/RoaringBitmap/src/main/java/org/roaringbitmap/RoaringBitmap.java
@@ -1338,100 +1338,6 @@ public class RoaringBitmap implements Cloneable, Serializable, Iterable<Integer>
     return andNot(x1, x2, (long) rangeStart, (long) rangeEnd);
   }
 
-  /*
-   * Handle the orNot operations for the remaining containers in the Bitmap.
-   * This is done by iterating over the remaining containers while treating the holes
-   * (from s2 till lastKey)
-   *
-   * For each iteration Two cases here:
-   * 1. either we have a existing container. In this case, we replace it by a full.
-   * 2. or there is no container. In this case, we insert a full one.
-   *
-   * Note that, at this stage, all the other bitmap containers were treated.
-   * That's why we only have to handle the .
-   */
-
-  private static char orNotHandleRemainingSelfContainers(
-          RoaringBitmap src, RoaringBitmap dest, int  pos1, int length1, char s2,
-          char lastKey, int lastSize, boolean inplace) {
-    final int insertionIncrement = inplace ? 1 : 0;
-    int destPos = inplace ? pos1 : dest.highLowContainer.size();
-
-    while (pos1 < length1 && s2 - lastKey <= 0) { // s2 <= lastKey
-      final char s1 = src.highLowContainer.getKeyAtIndex(pos1);
-      final int containerLast = (s2 == lastKey) ? lastSize : Util.maxLowBitAsInteger();
-      Container c2 = Container.rangeOfOnes(0, containerLast + 1);
-
-      if (s1 == s2) {
-        final Container c1 = src.highLowContainer.getContainerAtIndex(pos1);
-        // If we re not at the last container, just use the full container.
-        // Otherwise, compute an in-place or.
-        final Container c = (s2 == lastKey) ? (inplace ? c1.ior(c2) : c1.or(c2)) : c2;
-        if (!c.isEmpty()) {
-          if (destPos < dest.highLowContainer.size()) {
-            dest.highLowContainer.replaceKeyAndContainerAtIndex(destPos, s1, c);
-          } else {
-            dest.highLowContainer.insertNewKeyValueAt(destPos, s1, c);
-          }
-          pos1++;
-          s2++;
-          destPos++;
-        }
-      } else if (s1 - s2 > 0) { 
-        dest.highLowContainer.insertNewKeyValueAt(destPos, s2, c2);
-        pos1 += insertionIncrement;
-        length1 += insertionIncrement;
-        s2++;
-        destPos++;
-      } else { 
-        throw new IllegalStateException("This is a bug. Please report to github");
-      }
-    }
-    return s2;
-  }
-
-  /*
-   * Handle the orNot operations for the remaining containers in the other Bitmap.
-   * Two cases here:
-   * 1. either we have a hole. In this case, a full container should be appended.
-   * 2. or we have a container. an inplace orNot is applied and the result is appended.
-   *
-   * Note that, at this stage, all the own containers were treated.
-   * That's why we only have to append.
-   */
-  private static char orNotHandleRemainingOtherContainers(
-          final RoaringBitmap other, final RoaringBitmap dest, int pos2,
-          int length2, char s2, char lastKey, int lastSize) {
-    while (pos2 < length2 && s2 <= lastKey) {
-      final int containerLast = (s2 == lastKey) ? lastSize : Util.maxLowBitAsInteger();
-      if (s2 == other.highLowContainer.getKeyAtIndex(pos2)) {
-        final Container c2 = other.highLowContainer.getContainerAtIndex(pos2);
-        Container c = new RunContainer().iorNot(c2, containerLast + 1);
-        dest.highLowContainer.append(s2, c);
-        pos2++;
-      } else {
-        dest.highLowContainer.append(s2, Container.rangeOfOnes(0, containerLast + 1));
-      }
-      s2++;
-    }
-    return s2;
-  }
-
-  /*
-   * Handle the remaining holes.
-   * A full container should be appended for each key.
-   */
-  private static void orNotHandleRemainingHoles(
-          RoaringBitmap dest, char s2, char lastKey, int lastSize) {
-    while (s2 < lastKey) { 
-      dest.highLowContainer.append(s2, RunContainer.full());
-      s2++;
-    }
-    if (s2 == lastKey) {
-      dest.highLowContainer.append(s2, Container.rangeOfOnes(0, lastSize + 1));
-    }
-  }
-
   /**
    * In-place bitwise ORNOT operation. The current bitmap is modified.
    *
@@ -1441,155 +1347,69 @@ public class RoaringBitmap implements Cloneable, Serializable, Iterable<Integer>
    */
   public void orNot(final RoaringBitmap other, long rangeEnd) {
     rangeSanityCheck(0, rangeEnd);
-    if (rangeEnd == 0) {
-      clear();
-      return;
-    }
-
+    int maxKey = (int)(rangeEnd >>> 16);
+    int size = 0;
     int pos1 = 0, pos2 = 0;
-    int length1 = highLowContainer.size();
-    final int length2 = other.highLowContainer.size();
-
-    final char lastKey = Util.highbits(rangeEnd - 1);
-    final int lastSize = Util.lowbits(rangeEnd - 1);
-
-    char s2 = 0;
-    boolean loopedAtleastOnce = (length1 > 0 && length2 > 0);
-    while (pos1 < length1 && pos2 < length2
-            && s2 - lastKey <= 0) { // s2 <= lastKey
-      final char s1 = highLowContainer.getKeyAtIndex(pos1);
-      final int containerLast = (s2 == lastKey) ? lastSize : Util.maxLowBitAsInteger();
-
-      if (s1 == s2) {
-        final Container c1 = highLowContainer.getContainerAtIndex(pos1);
-        if (s2 == other.highLowContainer.getKeyAtIndex(pos2)) {
-          final Container c2 = other.highLowContainer.getContainerAtIndex(pos2);
-          final Container c = c1.iorNot(c2, containerLast + 1);
-          if (!c.isEmpty()) {
-            highLowContainer.replaceKeyAndContainerAtIndex(pos1, s1, c);
-            pos2++;
-          } else {
-            highLowContainer.removeAtIndex(pos1);
-          }
-        } else {
-          highLowContainer.replaceKeyAndContainerAtIndex(pos1, s1,
-                  Container.rangeOfOnes(0, containerLast + 1));
-        }
-        pos1++;
-        s2++;
-      } else if (s1 - s2 > 0) { 
-        if (s2 == other.highLowContainer.getKeyAtIndex(pos2)) {
-          final Container c2 = other.highLowContainer.getContainerAtIndex(pos2);
-          Container c = new RunContainer().iorNot(c2, containerLast + 1);
-          if (!c.isEmpty()) {
-            highLowContainer.insertNewKeyValueAt(pos1, s2, c);
-            pos2++;
-          }
-        } else {
-          highLowContainer.insertNewKeyValueAt(pos1, s2,
-                  Container.rangeOfOnes(0, containerLast + 1));
-        }
-        // Move forward because we inserted an element in the container.
-        pos1++;
-        length1++;
-        s2++;
-      } else { 
-        throw new IllegalStateException("This is a bug. Please report to github");
-      }
+    int length1 = highLowContainer.size(), length2 = other.highLowContainer.size();
+    int s1 = length1 > 0 ? highLowContainer.getKeyAtIndex(pos1) : maxKey + 1;
+    int s2 = length2 > 0 ? other.highLowContainer.getKeyAtIndex(pos2) : maxKey + 1;
+    int remainder = 0;
+    for (int i = highLowContainer.size - 1;
+         i >= 0 && highLowContainer.keys[i] > maxKey;
+         --i) {
+      ++remainder;
     }
-
-    // s2 == 0  means either that the bitmap is empty or that we wrapped around.
-    // In both cases, we want to stop
-    boolean loopHasWrapped = loopedAtleastOnce && (s2 == 0);
-    if (!loopHasWrapped && s2 - lastKey <= 0) { // s2 <= lastKey
-      char newS2;
-      if (pos1 < length1) {
-        //all the "other" arrays were treated. Handle self containers.
-        newS2 = orNotHandleRemainingSelfContainers(this, this, pos1, length1, s2,
-                lastKey, lastSize, true);
-
+    // it's almost certain that the bitmap will grow, so make a conservative overestimate,
+    // this avoids temporary allocation, and can trim afterwards
+    char[] newKeys = new char[maxKey + 1 + remainder];
+    Container[] newValues = new Container[maxKey + 1 + remainder];
+    for (int key = 0; key <= maxKey; ++key) {
+      if (key == s1 && key == s2) { // actually need to do an or not
+        newValues[size] = highLowContainer.getContainerAtIndex(pos1)
+                .iorNot(other.highLowContainer.getContainerAtIndex(pos2),
+                        key == maxKey ? (int)(rangeEnd & 0xFFFF) : 0x10000);
+        ++pos1;
+        ++pos2;
+        s1 = pos1 < length1 ? highLowContainer.getKeyAtIndex(pos1) : maxKey + 1;
+        s2 = pos2 < length2 ? other.highLowContainer.getKeyAtIndex(pos2) : maxKey + 1;
+      } else if (key == s1) { // or in a hole
+        newValues[size] = highLowContainer.getContainerAtIndex(pos1)
+                .ior(key == maxKey
+                  ? RunContainer.rangeOfOnes(0, (int)(rangeEnd & 0xFFFF))
+                  : RunContainer.full());
+        ++pos1;
+        s1 = pos1 < length1 ? highLowContainer.getKeyAtIndex(pos1) : maxKey + 1;
+      } else if (key == s2) { // insert the complement
+        newValues[size] = other.highLowContainer.getContainerAtIndex(pos2)
+                .not(0, key == maxKey ? (int)(rangeEnd & 0xFFFF) : 0x10000);
+        ++pos2;
+        s2 = pos2 < length2 ? other.highLowContainer.getKeyAtIndex(pos2) : maxKey + 1;
+      } else { // key missing from both
+        newValues[size] = key == maxKey
+                        ? RunContainer.rangeOfOnes(0, (int)(rangeEnd & 0xFFFF))
+                        : RunContainer.full();
+      }
+      // might have appended an empty container (rare case)
+      if (newValues[size].isEmpty()) {
+        newValues[size] = null;
       } else {
-        // all the original arrays were treated.
-        // We just need to iterate on the rest of the other arrays while handling holes.
-        newS2 = orNotHandleRemainingOtherContainers(other, this, pos2, length2, s2,
-                lastKey, lastSize);
-      }
-      // Check that we didn't wrap around
-      if (newS2 >= s2) {
-        orNotHandleRemainingHoles(this, newS2, lastKey, lastSize);
+        newKeys[size] = (char)key;
+        ++size;
       }
     }
-  }
-
-
-  private static RoaringBitmap doOrNot(
-          final RoaringBitmap rb1, final RoaringBitmap rb2, long rangeEnd) {
-    final RoaringBitmap answer = new RoaringBitmap();
-
-    int pos1 = 0, pos2 = 0;
-    int length1 = rb1.highLowContainer.size();
-    final int length2 = rb2.highLowContainer.size();
-
-    final char lastKey = Util.highbits(rangeEnd - 1);
-    final int lastSize = (Util.lowbits(rangeEnd - 1));
-
-    char s2 = 0;
-    boolean loopedAtleastOnce = (length1 > 0 && length2 > 0);
-    while (pos1 < length1 && pos2 < length2
-            && s2 - lastKey <= 0) { // s2 <= lastKey
-      final char s1 = rb1.highLowContainer.getKeyAtIndex(pos1);
-      final int containerLast = (s2 == lastKey) ? lastSize : Util.maxLowBitAsInteger();
-
-      if (s1 == s2) {
-        final Container c1 = rb1.highLowContainer.getContainerAtIndex(pos1);
-        if (s2 == rb2.highLowContainer.getKeyAtIndex(pos2)) {
-          final Container c2 = rb2.highLowContainer.getContainerAtIndex(pos2);
-          final Container c = c1.orNot(c2, containerLast + 1);
-          answer.highLowContainer.append(s1, c);
-          pos2++;
-        } else {
-          answer.highLowContainer.append(s1,
-                  Container.rangeOfOnes(0, containerLast + 1));
-        }
-        pos1++;
-        s2++;
-      } else if (s1 - s2 > 0) { 
-        if (s2 == rb2.highLowContainer.getKeyAtIndex(pos2)) {
-          final Container c2 = rb2.highLowContainer.getContainerAtIndex(pos2);
-          Container c = new RunContainer().orNot(c2, containerLast + 1);
-          answer.highLowContainer.append(s2, c);
-          pos2++;
-        } else {
-          answer.highLowContainer.append(s2,
-                  Container.rangeOfOnes(0, containerLast + 1));
-        }
-        s2++;
-      } else { 
-        throw new IllegalStateException("This is a bug. Please report to github");
+    // copy over everything which will remain without being complemented
+    if (remainder > 0) {
+      for (int i = 1; i <= remainder; ++i) {
+        int source = highLowContainer.size - i;
+        int target = newKeys.length - i;
+        newKeys[target] = highLowContainer.keys[source];
+        // no need to copy, the old storage will be lost anyway
+        newValues[target] = highLowContainer.values[source];
       }
     }
-
-    boolean loopHasWrapped = loopedAtleastOnce && (s2 == 0);
-    if (!loopHasWrapped && s2 - lastKey <= 0) { // s2 <= lastKey
-      char newS2;
-      if (pos1 < length1) {
-        //all the "other" arrays were treated. Handle self containers.
-        answer.highLowContainer.extendArray(lastKey + 1);
-        newS2 = orNotHandleRemainingSelfContainers(rb1, answer, pos1, length1, s2,
-                lastKey, lastSize, false);
-      } else {
-        // all the original arrays were treated.
-        // We just need to iterate on the rest of the other arrays while handling holes.
-        newS2 = orNotHandleRemainingOtherContainers(rb2, answer, pos2, length2, s2,
-                lastKey, lastSize);
-      }
-      // Check that we didnt wrap around
-      if (newS2 >= s2) {
-        orNotHandleRemainingHoles(answer, newS2, lastKey, lastSize);
-      }
-
-    }
-    return answer;
+    highLowContainer.keys = newKeys;
+    highLowContainer.values = newValues;
+    highLowContainer.size = size + remainder;
   }
 
   /**
@@ -1606,11 +1426,70 @@ public class RoaringBitmap implements Cloneable, Serializable, Iterable<Integer>
   public static RoaringBitmap orNot(
           final RoaringBitmap x1, final RoaringBitmap x2, long rangeEnd) {
     rangeSanityCheck(0, rangeEnd);
-
-    final RoaringBitmap rb1 = selectRangeWithoutCopy(x1, 0, rangeEnd);
-    final RoaringBitmap rb2 = selectRangeWithoutCopy(x2, 0, rangeEnd);
-
-    return doOrNot(rb1, rb2, rangeEnd);
+    int maxKey = (int)(rangeEnd >>> 16);
+    int size = 0;
+    int pos1 = 0, pos2 = 0;
+    int length1 = x1.highLowContainer.size(), length2 = x2.highLowContainer.size();
+    int s1 = length1 > 0 ? x1.highLowContainer.getKeyAtIndex(pos1) : maxKey + 1;
+    int s2 = length2 > 0 ? x2.highLowContainer.getKeyAtIndex(pos2) : maxKey + 1;
+    int remainder = 0;
+    for (int i = x1.highLowContainer.size - 1;
+         i >= 0 && x2.highLowContainer.keys[i] > maxKey;
+         --i) {
+      ++remainder;
+    }
+    // it's almost certain that the bitmap will grow, so make a conservative overestimate,
+    // this avoids temporary allocation, and can trim afterwards
+    char[] newKeys = new char[maxKey + 1 + remainder];
+    Container[] newValues = new Container[maxKey + 1 + remainder];
+    for (int key = 0; key <= maxKey; ++key) {
+      if (key == s1 && key == s2) { // actually need to do an or not
+        newValues[size] = x1.highLowContainer.getContainerAtIndex(pos1)
+                .orNot(x2.highLowContainer.getContainerAtIndex(pos2),
+                        key == maxKey ? (int)(rangeEnd & 0xFFFF) : 0x10000);
+        ++pos1;
+        ++pos2;
+        s1 = pos1 < length1 ? x1.highLowContainer.getKeyAtIndex(pos1) : maxKey + 1;
+        s2 = pos2 < length2 ? x2.highLowContainer.getKeyAtIndex(pos2) : maxKey + 1;
+      } else if (key == s1) { // or in a hole
+        newValues[size] = x1.highLowContainer.getContainerAtIndex(pos1)
+                .or(key == maxKey
+                        ? RunContainer.rangeOfOnes(0, (int)(rangeEnd & 0xFFFF))
+                        : RunContainer.full());
+        ++pos1;
+        s1 = pos1 < length1 ? x1.highLowContainer.getKeyAtIndex(pos1) : maxKey + 1;
+      } else if (key == s2) { // insert the complement
+        newValues[size] = x2.highLowContainer.getContainerAtIndex(pos2)
+                .not(0, key == maxKey ? (int)(rangeEnd & 0xFFFF) : 0x10000);
+        ++pos2;
+        s2 = pos2 < length2 ? x2.highLowContainer.getKeyAtIndex(pos2) : maxKey + 1;
+      } else { // key missing from both
+        newValues[size] = key == maxKey
+                ? RunContainer.rangeOfOnes(0, (int)(rangeEnd & 0xFFFF))
+                : RunContainer.full();
+      }
+      // might have appended an empty container (rare case)
+      if (newValues[size].isEmpty()) {
+        newValues[size] = null;
+      } else {
+        newKeys[size] = (char)key;
+        ++size;
+      }
+    }
+    // copy over everything which will remain without being complemented
+    if (remainder > 0) {
+      for (int i = 1; i <= remainder; ++i) {
+        int source = x1.highLowContainer.size - i;
+        int target = newKeys.length - i;
+        newKeys[target] = x1.highLowContainer.keys[source];
+        newValues[target] = x1.highLowContainer.values[source].clone();
+      }
+    }
+    RoaringBitmap result = new RoaringBitmap();
+    result.highLowContainer.keys = newKeys;
+    result.highLowContainer.values = newValues;
+    result.highLowContainer.size = size + remainder;
+    return result;
   }