--- a/fuzz-tests/build.gradle.kts
+++ b/fuzz-tests/build.gradle.kts
@@ -4,12 +4,12 @@ dependencies {
     implementation(project(":RoaringBitmap"))
     testImplementation("junit:junit:${deps["junit"]}")
     testImplementation("com.google.guava:guava:${deps["guava"]}")
-    testImplementation("com.fasterxml.jackson.core:jackson-databind:2.9.9")
+    testImplementation("com.fasterxml.jackson.core", "jackson-databind", "2.10.3")
 }
 
 tasks.test {
     // set the property on the CLI with -P or add to gradle.properties to enable tests
     if (!project.hasProperty("roaringbitmap.fuzz-tests")) {
-        exclude("**")
+       exclude("**")
     }
 }