Fix #376

1. add an emptiness test to RunContainer.and(RunContainer)
2. Special case check when `rangeEnd` is zero
3. stop empty containers from getting into the bitmap during orNot, resulting in whatever happens to be in an empty container being considered.
4. Add a comment to clarify the behaviour that the bitmap is truncated at `rangeEnd` (rightly or wrongly).

Does not change semantics