@lemire 2020-04-08
I think that the decision taken in https://github.com/RoaringBitmap/RoaringBitmap/issues/376 was not change the semantics. So this PR is presumably waiting to be updated.