@richardstartin 2020-04-08
I thought we would change the semantics, but I haven’t done it yet.