@lemire 2020-04-08
@richardstartin Thank you. I just wanted to leave a trace so that we know where we are at.