@richardstartin 2020-04-08
@lemire 

I have rewritten the method to observe the new semantics. The method adopts a heuristic that the array is almost guaranteed to grow, except in some degenerate cases, so always allocates a new array.