@lemire 2020-04-08
+1

You can merge this as far as I am concerned.