@richardstartin 2020-04-08
I'm putting something in to guard against `orNot` with full bitmaps blowing up, and will run some benchmarks to check it is better or any regression is < 10% on the status quo.