@richardstartin 2020-04-08
This is faster (Ubuntu 18, JDK11, microbenchmarks may not be representative of actual performance).

Before (3 data sets causes AIOOBE :-1: ):
```
Benchmark                                          (dataset)  Mode  Cnt       Score       Error  Units
RealDataBenchmarkOrNot.pairwiseOrNot           census-income  avgt    5   14287.869 ±   572.691  us/op
RealDataBenchmarkOrNot.pairwiseOrNot              census1881  avgt    5   11257.485 ±   511.279  us/op
RealDataBenchmarkOrNot.pairwiseOrNot           dimension_003  avgt    5  108631.069 ±  8026.821  us/op
RealDataBenchmarkOrNot.pairwiseOrNot           dimension_033  avgt    5   16226.913 ±   753.823  us/op
RealDataBenchmarkOrNot.pairwiseOrNot            uscensus2000  avgt    5   23824.822 ± 35955.458  us/op
RealDataBenchmarkOrNot.pairwiseOrNot         weather_sept_85  avgt    5   36488.019 ±  4292.528  us/op
RealDataBenchmarkOrNot.pairwiseOrNot      wikileaks-noquotes  avgt    5   14058.672 ±  1940.563  us/op
RealDataBenchmarkOrNot.pairwiseOrNot       census-income_srt  avgt    5   10181.575 ±   517.144  us/op
RealDataBenchmarkOrNot.pairwiseOrNot  wikileaks-noquotes_srt  avgt    5   10361.382 ±   548.670  us/op
```

After:

```
Benchmark                                          (dataset)  Mode  Cnt      Score      Error  Units
RealDataBenchmarkOrNot.pairwiseOrNot           census-income  avgt    5  15262.530 ± 2595.704  us/op
RealDataBenchmarkOrNot.pairwiseOrNot              census1881  avgt    5   6800.306 ±  323.475  us/op
RealDataBenchmarkOrNot.pairwiseOrNot           dimension_008  avgt    5  49573.086 ± 2517.147  us/op
RealDataBenchmarkOrNot.pairwiseOrNot           dimension_003  avgt    5  58861.510 ± 1395.904  us/op
RealDataBenchmarkOrNot.pairwiseOrNot           dimension_033  avgt    5  13281.648 ±  380.843  us/op
RealDataBenchmarkOrNot.pairwiseOrNot            uscensus2000  avgt    5   7706.333 ±   92.568  us/op
RealDataBenchmarkOrNot.pairwiseOrNot         weather_sept_85  avgt    5  37008.787 ± 7962.414  us/op
RealDataBenchmarkOrNot.pairwiseOrNot      wikileaks-noquotes  avgt    5  10999.235 ±  931.082  us/op
RealDataBenchmarkOrNot.pairwiseOrNot       census-income_srt  avgt    5   9431.580 ±  362.692  us/op
RealDataBenchmarkOrNot.pairwiseOrNot          census1881_srt  avgt    5   8562.199 ±  107.020  us/op
RealDataBenchmarkOrNot.pairwiseOrNot     weather_sept_85_srt  avgt    5  20442.076 ±  623.475  us/op
RealDataBenchmarkOrNot.pairwiseOrNot  wikileaks-noquotes_srt  avgt    5   7694.901 ±  627.728  us/op


```