@lemire 2020-04-08
Merging. I will issue a release as well.